+++ api-error/CHANGELOG.md
@@ -7,6 +7,36 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.5](https://github.com/centreon/rs-api-error/compare/api-error-v0.1.4...api-error-v0.1.5) - 2026-05-19
+
+### Added
+
+- add extended method behind axum feature flag  [release] ([#22](https://github.com/centreon/rs-api-error/pull/22))
+
+## [0.1.4](https://github.com/centreon/rs-api-error/compare/api-error-v0.1.3...api-error-v0.1.4) - 2026-05-11
+
+### Added
+
+- add error responder [release] ([#19](https://github.com/centreon/rs-api-error/pull/19))
+
+## [0.1.3](https://github.com/centreon/rs-api-error/compare/api-error-v0.1.1...api-error-v0.1.3) - 2026-04-21
+
+### Fixed
+
+- derive versioning [release] ([#17](https://github.com/centreon/rs-api-error/pull/17))
+- re-export transitive deps to avoid axum_core not found bug [release] ([#15](https://github.com/centreon/rs-api-error/pull/15))
+
+### Other
+
+- release v0.1.2 ([#14](https://github.com/centreon/rs-api-error/pull/14))
+- globally versionned crates [release] ([#13](https://github.com/centreon/rs-api-error/pull/13))
+
+## [0.1.2](https://github.com/centreon/rs-api-error/compare/api-error-v0.1.1...api-error-v0.1.2) - 2026-04-21
+
+### Other
+
+- globally versionned crates [release] ([#13](https://github.com/centreon/rs-api-error/pull/13))
+
 ## [0.1.1](https://github.com/centreon/rs-api-error/compare/api-error-v0.1.0...api-error-v0.1.1) - 2026-04-21
 
 ### Added